fix warnings.

+++ src/abacus/metrics.cpp
@@ -98,17 +98,17 @@ metrics::metrics(const metric_info* info, std::size_t count) :
 
     // The total bytes used for names
     assert(name_bytes <= std::numeric_limits<uint16_t>::max());
-    new (m_meta_data + 2) uint16_t(name_bytes);
+    new (m_meta_data + 2) uint16_t((uint16_t)name_bytes);
     // The total bytes used for descriptions
     assert(description_bytes <= std::numeric_limits<uint16_t>::max());
-    new (m_meta_data + 4) uint16_t(description_bytes);
+    new (m_meta_data + 4) uint16_t((uint16_t)description_bytes);
     // The total bytes used for units
     assert(unit_bytes <= std::numeric_limits<uint16_t>::max());
-    new (m_meta_data + 6) uint16_t(unit_bytes);
+    new (m_meta_data + 6) uint16_t((uint16_t)unit_bytes);
     // The number of 8-byte metric values (uint64_t, int64_t and double types)
-    new (m_meta_data + 8) uint16_t(m_info.eight_byte_metrics_count());
+    new (m_meta_data + 8) uint16_t((uint16_t)m_info.eight_byte_metrics_count());
     // The number of 1-byte metric values (bool type)
-    new (m_meta_data + 10) uint16_t(m_info.one_byte_metrics_count());
+    new (m_meta_data + 10) uint16_t((uint16_t)m_info.one_byte_metrics_count());
 
     // Write the name sizes into memory
     uint8_t* name_sizes_ptr = m_meta_data + detail::name_sizes_offset();
